requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
blis==1.3.0
build==1.2.2.post1
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
chroma-hnswlib==0.7.3
chromadb==0.5.0
click==8.2.1
cloudpathlib==0.21.1
cmudict==1.0.32
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
cryptography==45.0.4
cymem==2.0.11
dataclasses-json==0.6.7
distro==1.9.0
durationpy==0.10
emoji==2.14.1
exceptiongroup==1.3.0
fastapi==0.115.13
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.5.1
gensim==4.3.3
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.0
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
jellyfish==1.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
kubernetes==33.1.0
langchain==0.2.2
langchain-community==0.2.3
langchain-core==0.3.65
langchain-huggingface==0.3.0
langchain-openai==0.1.25
langchain-text-splitters==0.2.4
langcodes==3.5.0
langdetect==1.0.9
langsmith==0.3.45
language_data==1.3.0
lxml==5.4.0
marisa-trie==1.2.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mmh3==5.1.0
mpmath==1.3.0
multidict==6.5.0
murmurhash==1.0.13
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.8.1
numpy==1.26.4
oauthlib==3.3.0
onnxruntime==1.17.1
openai==1.55.3
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-instrumentation==0.55b1
opentelemetry-instrumentation-asgi==0.55b1
opentelemetry-instrumentation-fastapi==0.55b1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
opentelemetry-util-http==0.55b1
orjson==3.10.18
overrides==7.7.0
packaging==24.2
pillow==11.2.1
posthog==5.3.0
preshed==3.0.10
propcache==0.3.2
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.1
PyMuPDF==1.26.1
pypdf==5.6.0
pyphen==0.17.2
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2025.2.18
python-magic==0.4.27
PyYAML==6.0.2
rank-bm25==0.2.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.13.1
segtok==1.5.11
sentence-transformers==4.1.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
srsly==2.5.1
starlette==0.46.2
sympy==1.14.0
tabulate==0.9.0
tenacity==8.5.0
textstat==0.7.7
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.7.0
tokenizers==0.21.1
tomli==2.2.1
torch==2.7.1
tqdm==4.67.1
transformers==4.52.4
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
Unidecode==1.4.0
unstructured==0.14.4
unstructured-client==0.36.0
urllib3==2.5.0
uvicorn==0.34.3
wasabi==1.1.3
watchfiles==1.1.0
weasel==0.4.1
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
yake==0.4.8
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
